src/main/java/com/oocl/cultivation/ParkingBoy.java
package com.oocl.cultivation;

import com.oocl.cultivation.exception.UnrecognizedParkingTicketException;

public class ParkingBoy {
    ParkingLot parkingLot;

    public ParkingBoy(ParkingLot parkingLot) {
        this.parkingLot = parkingLot;
    }

    public ParkingTicket park(Car car) {
        if(!parkingLot.isParkingLotFull()){
            return parkingLot.park(car);
        }
        return null;
    }

    public Car fetch(ParkingTicket parkingTicket) throws UnrecognizedParkingTicketException {
        if(parkingTicket == null){
            throw new UnrecognizedParkingTicketException("Please provide your parking ticket.");
        }
        return parkingLot.fetch(parkingTicket);
    }
}
